Fixed a bug in DoubleHash, and QPHashTable where there was integer overflow in the step calculation (step as in the argument for hashByStepFromBase).
hashByStepFromBase now uses the normal remainder operator instead of floorMod, causing it to require step>=0 as a precondition.
AQPHashTable now always passes a positive step to the method,
accordingly.

The currSize field in OAHashTable is now updated correctly.

+++ src/DoubleHashTable.java
@@ -10,7 +10,10 @@ public class DoubleHashTable extends OAHashTable {
 
     @Override
     public int Hash(long x, int i) {
-        return hashByStepFromBase(x,i*(stepHash.Hash(x)+1));
+        return hashByStepFromBase(x, (long)i*(stepHash.Hash(x)+1));
+        // long cast is necessary because the expression can overflow int boundaries
+        // (e.g. when both i and the step equal m-1, and m-1 has more than 16 bits,
+        // causing (m-1)*(m-1) to overflow past 32 bits).
     }
 
 }
